# fronted/OPENAPI/ExamSystemApi.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ExamSystemApi",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/v1/messages/{page}/{size}": {
      "get": {
        "summary": "findAll",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultMessage"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": {
                        "id": 0,
                        "temp_id": 0,
                        "title": "",
                        "content": "",
                        "time": "",
                        "replays": [
                          {
                            "messageId": 0,
                            "replayId": 0,
                            "replay": "",
                            "replayTime": ""
                          }
                        ]
                      }
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/message/{id}": {
      "get": {
        "summary": "findById",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "根据Id查询",
                      "data": {
                        "id": 0,
                        "temp_id": 0,
                        "title": "",
                        "content": "",
                        "time": "",
                        "replays": [
                          {
                            "messageId": 0,
                            "replayId": 0,
                            "replay": "",
                            "replayTime": ""
                          }
                        ]
                      }
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "delete",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": 0
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/message": {
      "post": {
        "summary": "add",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Message",
                "description": "返回的结果包装在value中，value可以是单个对象"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "添加成功",
                      "data": 0
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/v1/exams/{page}/{size}": {
      "get": {
        "summary": "findAll",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": {}
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/exam/{examCode}": {
      "get": {
        "summary": "findById",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "examCode",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "请求成功！",
                      "data": {
                        "examCode": 0,
                        "description": "",
                        "source": "",
                        "paperId": 0,
                        "examDate": "",
                        "totalTime": 0,
                        "grade": "",
                        "term": "",
                        "major": "",
                        "institute": "",
                        "totalScore": 0,
                        "type": "",
                        "tips": ""
                      }
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "deleteById",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "examCode",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "删除成功",
                      "data": 0
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/exam": {
      "put": {
        "summary": "update",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamManage",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "更新成功",
                      "data": 0
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "add",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamManage",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 400,
                      "message": "添加失败",
                      "data": 0
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "错误码，表示一种错误类型\n请求成功，状态码为200"
          },
          "message": {
            "type": "string",
            "description": "对错误代码的详细解释"
          },
          "data": {
            "type": "integer",
            "description": "返回的结果包装在value中，value可以是单个对象"
          }
        }
      },
      "Replay": {
        "type": "object",
        "properties": {
          "messageId": {
            "type": "integer",
            "description": ""
          },
          "replayId": {
            "type": "integer",
            "description": ""
          },
          "replay": {
            "type": "string",
            "description": ""
          },
          "replayTime": {
            "type": "string",
            "description": ""
          }
        }
      },
...
      "Message": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": ""
          },
          "temp_id": {
            "type": "integer",
            "description": "解决id为null创建的一个临时id"
          },
          "title": {
            "type": "string",
            "description": ""
          },
          "content": {
            "type": "string",
            "description": ""
          },
          "time": {
            "type": "string",
            "description": ""
          },
          "replays": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Replay",
              "description": "fun.timu.exam.entity.Replay"
            },
            "description": "一对多关系，评论信息"
          }
        }
      },
...
      "ApiResultMessage": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "错误码，表示一种错误类型\n请求成功，状态码为200"
          },
          "message": {
            "type": "string",
            "description": "对错误代码的详细解释"
          },
          "data": {
            "$ref": "#/components/schemas/Message",
            "description": "返回的结果包装在value中，value可以是单个对象"
          }
        }
      },
      "ExamManage": {
        "type": "object",
        "properties": {
          "examCode": {
            "type": "integer",
            "description": ""
          },
          "description": {
            "type": "string",
            "description": ""
          },
          "source": {
            "type": "string",
            "description": ""
          },
          "paperId": {
            "type": "integer",
            "description": ""
          },
          "examDate": {
            "type": "string",
            "description": ""
          },
          "totalTime": {
            "type": "integer",
            "description": ""
          },
          "grade": {
            "type": "string",
            "description": ""
          },
          "term": {
            "type": "string",
            "description": ""
          },
          "major": {
            "type": "string",
            "description": ""
          },
          "institute": {
            "type": "string",
            "description": ""
          },
          "totalScore": {
            "type": "integer",
            "description": ""
          },
          "type": {
            "type": "string",
            "description": ""
          },
          "tips": {
            "type": "string",
            "description": ""
          }
        }
      },
...
    },
    "securitySchemes": {}
  },
  "servers": []
}