# fronted/src/generated/models/Replay.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Replay = {
    messageId?: number;
    replayId?: number;
    replay?: string;
    replayTime?: string;
};